Minor cleanups
TODO: cell_generator, sandbox and cell_template need more cleanups.
