+++ lib/src/kernel.dart
@@ -9,6 +9,7 @@ import 'dart:isolate';
 import 'package:dart_repl/src/package_utils.dart';
 import 'package:dart_repl/src/sandbox_isolate.dart';
 import 'package:dart_repl/src/cell_type.dart';
+import 'package:dart_repl_sandbox/builtin_commands/messages.dart';
 import 'package:dart_repl_sandbox/data_queue.dart';
 import 'package:dart_repl_sandbox/isolate_messages.dart';
 import 'package:dart_repl_sandbox/message_queue.dart';
