+++ sandbox/lib/builtin_commands/api.dart
@@ -1,6 +1,10 @@
+// Copyright (c) 2017, Andreas 'blackhc' Kirsch. All rights reserved. Use of
+// this source code is governed by a BSD-style license that can be found in the
+// LICENSE file.
+
 import 'dart:async';
 import 'dart:isolate';
-import 'package:dart_repl_sandbox/isolate_messages.dart';
+import 'package:dart_repl_sandbox/builtin_commands/messages.dart';
 import 'package:dart_repl_sandbox/message_channel.dart';
 
 MessageSender sandboxRequestSender;
