+++ sandbox/lib/cell_environment.dart
@@ -1,6 +1,7 @@
 // Copyright (c) 2016, Andreas 'blackhc' Kirsch. All rights reserved. Use of
 // this source code is governed by a BSD-style license that can be found in the
 // LICENSE file.
+
 import 'package:dart_repl_sandbox/cell.dart' as cell;
 import 'scope.dart';
