+++ sandbox/lib/isolate_messages.dart
@@ -65,40 +65,3 @@ class CellResult extends Message<CellResult> {
 
 final CellReplyConverters =
     new MessageConverter({CellResult: CellResult.fromRawMessage});
-
-class ImportLibraryRequest extends Message<ImportLibraryRequest> {
-  final String libraryPath;
-
-  ImportLibraryRequest(this.libraryPath);
-
-  static ImportLibraryRequest fromRawMessage(Object simpleFormat) =>
-      new ImportLibraryRequest(simpleFormat as String);
-
-  @override
-  Object toRawData() => libraryPath;
-}
-
-class LoadPackageRequest extends Message<LoadPackageRequest> {
-  final String packageName;
-  final String packageConfigUri;
-
-  LoadPackageRequest(this.packageName, this.packageConfigUri);
-
-  static LoadPackageRequest fromRawMessage(Object simpleFormat) {
-    final dict = simpleFormat as Map<String, String>;
-    return new LoadPackageRequest(dict['name'], dict['configUri']);
-  }
-
-  @override
-  Object toRawData() => {'name': packageName, 'configUri': packageConfigUri};
-}
-
-class ExitRequest extends Message<ExitRequest> {
-  static ExitRequest fromRawMessage(Object simpleFormat) => new ExitRequest();
-}
-
-final SandboxRequestQueueConverters = new MessageConverter({
-  ImportLibraryRequest: ImportLibraryRequest.fromRawMessage,
-  LoadPackageRequest: LoadPackageRequest.fromRawMessage,
-  ExitRequest: ExitRequest.fromRawMessage
-});
