+++ sandbox/lib/message_channel.dart
@@ -1,3 +1,7 @@
+// Copyright (c) 2017, Andreas 'blackhc' Kirsch. All rights reserved. Use of
+// this source code is governed by a BSD-style license that can be found in the
+// LICENSE file.
+
 import 'dart:async';
 import 'dart:isolate';
 import 'package:dart_repl_sandbox/data_queue.dart';
